Aggiunto il package http per gestire l'invio dei dati ad un realtime database

+++ lib/widgets/user_product_item.dart
@@ -37,8 +37,31 @@ class UserProductItem extends StatelessWidget {
               ),
               IconButton(
                 onPressed: () {
-                  Provider.of<ProductsProvider>(context, listen: false)
-                      .deleteProduct(id as String);
+                  showDialog(
+                    context: context,
+                    builder: (context) => AlertDialog(
+                      title: const Text('Delete the product?'),
+                      content: const Text(
+                          'Are you sure to delete definitely the product ?'),
+                      actions: [
+                        TextButton(
+                          onPressed: () {
+                            Navigator.of(context).pop(false);
+                          },
+                          child: const Text('Cancel'),
+                        ),
+                        TextButton(
+                          onPressed: () {
+                            Provider.of<ProductsProvider>(context,
+                                    listen: false)
+                                .deleteProduct(id as String);
+                            Navigator.of(context).pop(true);
+                          },
+                          child: const Text('Continue'),
+                        ),
+                      ],
+                    ),
+                  );
                 },
                 icon: const Icon(Icons.delete),
                 color: Theme.of(context).errorColor,
